- Set Compiler & Debugging levels to Extended and resolved most of the compiler warnings about casting and data type conversions for files DAQLab.c and LaserScanning.c - Changed number of elements in a waveform to be size_t instead of uInt64, since malloc can allocate only up to size_t_max bytes. - Not all variables are initialized on their declaration, this will take more time to correct. - The remaining files have not yet been corrected for the compiler warnings.

+++ Modules/VUPhotonCtr/HW_VUPC.c
@@ -254,9 +254,9 @@ int ReadBuffer(int bufsize)
 	char*					errMsg			= NULL;  
 	unsigned short int* 	Samplebuffer	= NULL;   
 	void*     				pmtdataptr		= NULL;
-	int 					numpixels;
-	int 					numshorts;
-	int 					ndatapoints;
+	size_t 					numpixels;
+	size_t 					numshorts;
+	size_t 					ndatapoints;
 	double 					refSamplingRate	= 1000;
 	long 					errcode;
 	unsigned long 			statreg;
